Desktop/django/xtension/requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
djongo==1.2.32
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-language==1.2.0
googleapis-common-protos==1.5.9
grpcio==1.20.0
gunicorn==19.9.0
idna==2.6
psycopg2==2.8.2
pytz==2019.1
requests==2.18.4
sqlparse==0.3.0
urllib3==1.22
whitenoise==4.1.2
